refactor: HttpResponse로 분리
